refactor: NotificationService 반환값 변경
